# docs/arch/API_Contract_v1.0.yaml
openapi: "3.0.3"
info:
  title: Learning OS API
  description: AI-independent learning execution system — REST API contract v1.0
  version: "1.0.0"
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/users/me:
    get:
      summary: Get current user profile
      operationId: getCurrentUser
      tags: [User]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "u1"
                nickname: "Learner"
                avatar: "https://example.com/avatar.png"
                xp: 350
                level: 4
                streak: 7
                created_at: "2026-06-01T00:00:00Z"
                updated_at: "2026-06-08T15:00:00Z"

    patch:
      summary: Update user profile
      operationId: updateCurrentUser
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              nickname: "Python Master"
              avatar: "https://example.com/new-avatar.png"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
              example:
                id: "u1"
                nickname: "Python Master"
                avatar: "https://example.com/new-avatar.png"
                xp: 350
                level: 4
                streak: 7
                created_at: "2026-06-01T00:00:00Z"
                updated_at: "2026-06-08T15:05:00Z"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

# ──────────────────────────────────────────────
#  COMPONENTS / SCHEMAS
# ──────────────────────────────────────────────
components:
  schemas:
...
    UpdateUserRequest:
      type: object
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 50
        avatar:
          type: string
          format: uri
          maxLength: 500

    # ── Response Schemas ──
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        avatar:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Error Schemas ──
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string
...